+++ src/python/arrow_group.py
@@ -1,4 +1,4 @@
-#arrow_group.py
+"""Methods for turning the generators of a group into the full group and getting the full arrow group"""
 
 #a_group take the set of translations of the lattice and the set of
 #rotations of the lattice paired with their effect on the arrows of
@@ -6,9 +6,15 @@
 #rotations and pairs them with the combination of th arraws. If the
 #resultant operation is not in the group it then adds it to the group.
 def a_group(trans,rots):
-    """
-    This subroutine combines that translations of the lattice with the
-    rotatians of the lattice.
+    """This subroutine combines that translations of the lattice with the
+      rotatians of the lattice.
+
+      :arg trans: a 2D integer array where each row is an translation
+      of the lattice
+      :arg rots: a 3D integer array. Each row's first entry is the
+      site permutations and the second entry is the arrow
+      permutations.
+
     """
     groupi = []
     for i in trans:
@@ -24,9 +30,12 @@ def a_group(trans,rots):
     return(groupi)
 
 def group(gen):
-    """
-    This subroutine takes the generators of the group then uses them
-    to form the entire group.
+    """This subroutine takes the generators of the group then uses them
+      to form the entire group.
+
+      :arg gen: a 2D integer array of the generators of the group
+      where each row is a seperate generator
+
     """
     # q is a counter to track how many group elements we've found
     q = 1