+++ src/python/binomial_calculator.py
@@ -1,9 +1,13 @@
-#binomial_calculator.py
+"""Methods used for calculating binomial coefficients"""
 
 #This method finds the factorial of a given number(num).
 #The method calls for an integer input num
 #The method returns the factorial(represented as fact in the method)
 def factorial(num):
+        """Finds the factorial of the input integer.
+
+          :arg num: an integer
+        """
         #If the number provided is zero then the factorial is 1
 	if num == 0:
 		fact = 1
@@ -21,6 +25,12 @@ def factorial(num):
 #choose r The method calls for two integer values n and r and the
 #method factorial The method returns the binomial factorial as result
 def binomial_coefficient(n,r):
+        """Finds the binomial coefficient, n choose r, for a given set of
+          integers.
+
+          :arg n: an integer
+          :arg r: an integer
+        """
         #If r is less than zero then the binomial coefficient is zero
 	if r < 0:
 		result = 0